Use route helpers to reduce duplication in route definition files
