+++ src/internal/modules/address-entry/routes.js
@@ -1,6 +1,7 @@
 'use strict';
 
 const { createPreHandler } = require('shared/lib/pre-handlers/forms');
+const { createRoutePair } = require('shared/lib/route-helpers');
 
 const controller = require('./controller');
 const preHandlers = require('./pre-handlers');
@@ -63,10 +64,8 @@ module.exports = {
     }
   },
 
-  getManualAddressEntry: {
-    method: 'GET',
+  ...createRoutePair(controller, 'manualAddressEntry', {
     path: '/address-entry/{key}/manual-entry',
-    handler: controller.getManualAddressEntry,
     options: {
       auth: {
         scope: allowedScopes
@@ -83,34 +82,10 @@ module.exports = {
         method: createPreHandler(forms.manualAddressEntry), assign: 'form'
       }]
     }
-  },
+  }),
 
-  postManualAddressEntry: {
-    method: 'POST',
-    path: '/address-entry/{key}/manual-entry',
-    handler: controller.postManualAddressEntry,
-    options: {
-      auth: {
-        scope: allowedScopes
-      },
-      description: 'Enter address manually',
-      plugins: {
-        viewContext: {
-          activeNavLink: 'notifications'
-        }
-      },
-      pre: [{
-        method: preHandlers.getSessionData, assign: 'sessionData'
-      }, {
-        method: createPreHandler(forms.manualAddressEntry), assign: 'form'
-      }]
-    }
-  },
-
-  getSelectCompanyAddress: {
-    method: 'GET',
+  ...createRoutePair(controller, 'selectCompanyAddress', {
     path: '/address-entry/{key}/select-company-address',
-    handler: controller.getSelectCompanyAddress,
     options: {
       auth: {
         scope: allowedScopes
@@ -131,38 +106,10 @@ module.exports = {
         method: createPreHandler(forms.selectCompanyAddress), assign: 'form'
       }]
     }
-  },
+  }),
 
-  postSelectCompanyAddress: {
-    method: 'POST',
-    path: '/address-entry/{key}/select-company-address',
-    handler: controller.postSelectCompanyAddress,
-    options: {
-      auth: {
-        scope: allowedScopes
-      },
-      description: 'Select an existing company address',
-      plugins: {
-        viewContext: {
-          activeNavLink: 'notifications'
-        }
-      },
-      pre: [{
-        method: preHandlers.getSessionData, assign: 'sessionData'
-      }, {
-        method: preHandlers.getCompanyAddresses, assign: 'addresses'
-      }, {
-        method: preHandlers.getCompany, assign: 'company'
-      }, {
-        method: createPreHandler(forms.selectCompanyAddress), assign: 'form'
-      }]
-    }
-  },
-
-  getUseRegisteredOfficeAddress: {
-    method: 'get',
+  ...createRoutePair(controller, `useRegisteredAddress`, {
     path: '/address-entry/{key}/use-registered-address',
-    handler: controller.getUseRegisteredAddress,
     options: {
       auth: {
         scope: allowedScopes
@@ -183,31 +130,6 @@ module.exports = {
         method: createPreHandler(forms.useRegisteredAddress), assign: 'form'
       }]
     }
-  },
+  })
 
-  postUseRegisteredOfficeAddress: {
-    method: 'post',
-    path: '/address-entry/{key}/use-registered-address',
-    handler: controller.postUseRegisteredAddress,
-    options: {
-      auth: {
-        scope: allowedScopes
-      },
-      description: 'Whether to use registered company address',
-      plugins: {
-        viewContext: {
-          activeNavLink: 'notifications'
-        }
-      },
-      pre: [{
-        method: preHandlers.getSessionData, assign: 'sessionData'
-      },
-      {
-        method: preHandlers.getCompaniesHouseCompany, assign: 'company'
-      },
-      {
-        method: createPreHandler(forms.useRegisteredAddress), assign: 'form'
-      }]
-    }
-  }
 };
