+++ src/internal/modules/returns-notifications/routes/paper-forms.js
@@ -7,10 +7,10 @@ const preHandlers = require('../pre-handlers');
 const eventPreHandlers = require('shared/lib/pre-handlers/events');
 const constants = require('../../../lib/constants');
 const { returns } = constants.scope;
-const { createFormRoutes } = require('../lib/route-helpers');
+const { createFormRoutes } = require('shared/lib/route-helpers');
 
 const formRoutes = {
-  getEnterLicenceNumber: {
+  enterLicenceNumber: {
     path: '/returns-notifications/forms',
     config: {
       auth: {
@@ -25,7 +25,7 @@ const formRoutes = {
     }
   },
 
-  getCheckAnswers: {
+  checkAnswers: {
     path: '/returns-notifications/check-answers',
     config: {
       auth: {
@@ -43,7 +43,7 @@ const formRoutes = {
     }
   },
 
-  getSelectReturns: {
+  selectReturns: {
     path: '/returns-notifications/{documentId}/select-returns',
     config: {
       auth: {
@@ -66,7 +66,7 @@ const formRoutes = {
     }
   },
 
-  getSelectAddress: {
+  selectAddress: {
     path: '/returns-notifications/{documentId}/select-address',
     config: {
       auth: {
@@ -89,7 +89,7 @@ const formRoutes = {
     }
   },
 
-  getRecipient: {
+  recipient: {
     path: '/returns-notifications/{documentId}/recipient',
     config: {
       auth: {
@@ -112,7 +112,7 @@ const formRoutes = {
     }
   },
 
-  getSelectLicenceHolders: {
+  selectLicenceHolders: {
     path: '/returns-notifications/select-licence-holders',
     config: {
       auth: {
